Adding deps.py.
This is a simple tool for compiling a list of unique ependencies based on gcc's .d file output.
This list can be used as a file list for ctags to build a tags file that includes all files in a build rather than all files in a source tree including those that aren't ini the build.